# src/swagger/api-swagger-v2.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Retrieves and parses information from the google maps api.",
    "version" : "0.1",
    "title" : "gmaps-adapter"
  },
  "paths" : {
    "/api/v1/directions" : {
      "post" : {
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model1"
          },
          "x-examples" : {
            "application/json" : "{\n\t\"origin\": \"Miami Lakes, florida\",\n\t\"destination\": \"Miami Beach, Fl\",\n\t\"departureTime\": \"15217600000\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok",
            "schema" : {
              "$ref" : "#/definitions/directionsRes"
            }
          }
        }
      }
    },
    "/api/v1/geocode" : {
      "post" : {
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model0"
          },
          "x-examples" : {
            "application/json" : "{\n\t\"address\": \"Miami Lakes, florida\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ok",
            "schema" : {
              "$ref" : "#/definitions/geocodeRes"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Model0" : {
      "properties" : {
        "address" : {
          "type" : "string"
        }
      },
      "example" : {
        "address" : "address"
      }
    },
    "Model1" : {
      "properties" : {
        "origin" : {
          "type" : "string"
        },
        "destination" : {
          "type" : "string"
        },
        "departureTime" : {
          "type" : "string"
        }
      },
      "example" : {
        "departureTime" : "departureTime",
        "origin" : "origin",
        "destination" : "destination"
      }
    },
    "directionsRes" : {
      "properties" : {
        "duration" : {
          "type" : "integer"
        },
        "distance" : {
          "type" : "integer"
        }
      },
      "example" : {
        "duration" : 0,
        "distance" : 6
      }
    },
    "geocodeRes" : {
      "properties" : {
        "latitude" : {
          "type" : "number"
        },
        "longitude" : {
          "type" : "number"
        }
      },
      "example" : {
        "latitude" : 0.80082819046101150206595775671303272247314453125,
        "longitude" : 6.02745618307040320615897144307382404804229736328125
      }
    }
  }
}